+++ Modules/Intent.Modules.AspNetCore.Scalar/release-notes.md
@@ -1,6 +1,12 @@
 ### Version 1.0.2
 
 - Improvement: Updated NuGet package versions.
+- Fixed: Aligned with `Intent.AspNetCore.Swashbuckle` so that Schema identifier generation to ensure generated swagger schema is OpenAPI compliant.
+
+> ⚠️ NOTE
+>
+> When using simple schema identifiers, generic types will use `Of` and `And` to separate generic types and parameters.
+> When using full-namespace identifiers, the identifier will use `_Of_` and `_And_` to separate generic types and parameters.
 
 ### Version 1.0.1